pyproject.toml
[project]
name = "aqi-probability-prediction"
version = "0.1.0"
description = "AQI Probability Prediction Pipeline"
readme = "README.md"
requires-python = "==3.10.11"

dependencies = [
  "numpy==1.26.4",
  "pandas==2.2.2",
  "optuna==4.3.0",
  "matplotlib==3.10.0",
  "scipy==1.12.0",
  "tqdm==4.67.1",
  "imageio==2.36.1",
  "pillow==11.0.0",
  "pip==25.1.1",
  "ipykernel==6.29.5",
  "widgetsnbextension==4.0.14",
    "ipywidgets==8.1.7",
    "pre-commit==4.2.0",

]

[project.optional-dependencies]
cuda = [
  "torch==2.6.0+cu126", 
  "torchvision==0.21.0"
]
metal = [
  "torch",
  "torchvision",
  "torchaudio"
]


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"